feat(S1.2.2.3): add driver register validations
